New Event crud

+++ src/main/java/com/uq/jokievents/service/EventService.java
@@ -1,8 +1,11 @@
 package com.uq.jokievents.service;
 
+import com.uq.jokievents.model.Client;
 import com.uq.jokievents.model.Event;
 import com.uq.jokievents.repository.EventRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,8 +22,13 @@ public class EventService {
      *
      * @return a list of all event objects in the db
      */
-    public List<Event> findAll() {
-        return eventRepository.findAll();
+    public ResponseEntity<?> findAll() {
+        try {
+            List<Event> events = eventRepository.findAll();
+            return new ResponseEntity<>(events, HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>("Failed event request", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     /**
@@ -29,8 +37,17 @@ public class EventService {
      * @param id the identifier of the event
      * @return an Optional containing the event if found, empty if not
      */
-    public Optional<Event> findById(String id) {
-        return eventRepository.findById(id);
+    public ResponseEntity<?> findById(String id) {
+        try {
+            Optional<Event> event = eventRepository.findById(id);
+            if (event.isPresent()) {
+                return new ResponseEntity<>(event.get(), HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("Event not found", HttpStatus.NOT_FOUND);
+            }
+        } catch (Exception e) {
+            return new ResponseEntity<>("Failed event request", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     /**
@@ -39,8 +56,35 @@ public class EventService {
      * @param event the event object to be saver or updated
      * @return the saved or updated event object
      */
-    public Event save(Event event) {
-        return eventRepository.save(event);
+    public ResponseEntity<?> create(Event event) {
+        try {
+            Event createdEvent = eventRepository.save(event);
+            return new ResponseEntity<>(createdEvent, HttpStatus.CREATED);
+        } catch (Exception e) {
+            return new ResponseEntity<>("Failed to create event", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    /**
+     * Update an existing client by id
+     *
+     * @param id the identifierof the event to be update
+     * @param event the updated event object
+     * @return a ResponseEntity containing the updated event object and an HTTP status
+     */
+    public ResponseEntity<?> update(String id, Event event) {
+        try {
+            Optional<Event> existingEvent = eventRepository.findById(id);
+            if (existingEvent.isPresent()) {
+                event.setId(id);
+                Event updatedEvent = eventRepository.save(event);
+                return new ResponseEntity<>(updatedEvent, HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("Event not found", HttpStatus.NOT_FOUND);
+            }
+        } catch (Exception e) {
+            return new ResponseEntity<>("Failed to update event", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     /**
@@ -48,8 +92,18 @@ public class EventService {
      *
      * @param id the identifier of the event object
      */
-    public void deleteById(String id) {
-        eventRepository.deleteById(id);
+    public ResponseEntity<?> deleteById(String id) {
+        try {
+            Optional<Event> existingEvent = eventRepository.findById(id);
+            if (existingEvent.isPresent()) {
+                eventRepository.deleteById(id);
+                return new ResponseEntity<>("Event deleted", HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("Event not found", HttpStatus.NOT_FOUND);
+            }
+        } catch (Exception e) {
+            return new ResponseEntity<>("Failed to delete event", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
 }
